Extract classes to files
to improve readability. Rename `analyzer` directory to `src` to avoid
import issues in tests.
